--- javascript/react/docs/markdown/react.stepcomponentprops.index.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepComponentProps](./react.stepcomponentprops.md) &gt; [index](./react.stepcomponentprops.index.md)
-
-## StepComponentProps.index property
-
-The step index
-
-<b>Signature:</b>
-
-```typescript
-index: number;
-```
