--- javascript/react/docs/markdown/react.stepcomponentprops.inputtablelabel.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepComponentProps](./react.stepcomponentprops.md) &gt; [inputTableLabel](./react.stepcomponentprops.inputtablelabel.md)
-
-## StepComponentProps.inputTableLabel property
-
-Label for the input-table dropdown
-
-<b>Signature:</b>
-
-```typescript
-inputTableLabel?: string;
-```
