--- javascript/react/docs/markdown/react.stepcomponentprops.md
@@ -1,32 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepComponentProps](./react.stepcomponentprops.md)
-
-## StepComponentProps interface
-
-<b>Signature:</b>
-
-```typescript
-export interface StepComponentProps 
-```
-
-## Properties
-
-|  Property | Modifiers | Type | Description |
-|  --- | --- | --- | --- |
-|  [hideInput?](./react.stepcomponentprops.hideinput.md) |  | boolean | <i>(Optional)</i> Indicates that the input table should be hidden or else shown and editable by the user. It may be desirable to hide this if the modal is launched directly from a table, which would make display redundant. |
-|  [hideInputColumn?](./react.stepcomponentprops.hideinputcolumn.md) |  | boolean | <i>(Optional)</i> Indicates that the input column should be hidden or else shown and editable by the user. It may be desirable to hide this if the transform is expected to use the selected column when the modal is launched directly from a table. |
-|  [hideOutput?](./react.stepcomponentprops.hideoutput.md) |  | boolean | <i>(Optional)</i> Indicates that the output table should be hidden or else shown and editable by the user. It may be desirable to hide this if the transform is expected to do an inline replacement of the input table. |
-|  [index](./react.stepcomponentprops.index.md) |  | number | The step index |
-|  [inputColumnLabel?](./react.stepcomponentprops.inputcolumnlabel.md) |  | string | <i>(Optional)</i> Label for the input table dropdown |
-|  [inputTableLabel?](./react.stepcomponentprops.inputtablelabel.md) |  | string | <i>(Optional)</i> Label for the input-table dropdown |
-|  [metadata?](./react.stepcomponentprops.metadata.md) |  | TableMetadata | <i>(Optional)</i> TableMetadata |
-|  [onChange](./react.stepcomponentprops.onchange.md) |  | (step: Step, index: number) =&gt; void | The step onchange handler |
-|  [onChangeOutput](./react.stepcomponentprops.onchangeoutput.md) |  | (value: string \| undefined) =&gt; void | Event handler for when the output table name changes |
-|  [output?](./react.stepcomponentprops.output.md) |  | string | <i>(Optional)</i> The output table name |
-|  [outputColumnLabel?](./react.stepcomponentprops.outputcolumnlabel.md) |  | string | <i>(Optional)</i> The label to use for the output column field |
-|  [outputTableDisabled?](./react.stepcomponentprops.outputtabledisabled.md) |  | boolean | <i>(Optional)</i> Whether the output table name is disabled |
-|  [outputTableLabel?](./react.stepcomponentprops.outputtablelabel.md) |  | string | <i>(Optional)</i> The label to use for the output field |
-|  [step](./react.stepcomponentprops.step.md) |  | Step | The processing step |
-|  [workflow](./react.stepcomponentprops.workflow.md) |  | Workflow | The workflow instance |
-
