--- javascript/react/docs/markdown/react.stepcomponentprops.metadata.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepComponentProps](./react.stepcomponentprops.md) &gt; [metadata](./react.stepcomponentprops.metadata.md)
-
-## StepComponentProps.metadata property
-
-TableMetadata
-
-<b>Signature:</b>
-
-```typescript
-metadata?: TableMetadata;
-```
