--- javascript/react/docs/markdown/react.stepcomponentprops.onchange.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepComponentProps](./react.stepcomponentprops.md) &gt; [onChange](./react.stepcomponentprops.onchange.md)
-
-## StepComponentProps.onChange property
-
-The step onchange handler
-
-<b>Signature:</b>
-
-```typescript
-onChange: (step: Step, index: number) => void;
-```
