--- javascript/react/docs/markdown/react.stepcomponentprops.output.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepComponentProps](./react.stepcomponentprops.md) &gt; [output](./react.stepcomponentprops.output.md)
-
-## StepComponentProps.output property
-
-The output table name
-
-<b>Signature:</b>
-
-```typescript
-output?: string;
-```
