--- javascript/react/docs/markdown/react.stepcomponentprops.outputcolumnlabel.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepComponentProps](./react.stepcomponentprops.md) &gt; [outputColumnLabel](./react.stepcomponentprops.outputcolumnlabel.md)
-
-## StepComponentProps.outputColumnLabel property
-
-The label to use for the output column field
-
-<b>Signature:</b>
-
-```typescript
-outputColumnLabel?: string;
-```
