--- javascript/react/docs/markdown/react.stepcomponentprops.outputtabledisabled.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepComponentProps](./react.stepcomponentprops.md) &gt; [outputTableDisabled](./react.stepcomponentprops.outputtabledisabled.md)
-
-## StepComponentProps.outputTableDisabled property
-
-Whether the output table name is disabled
-
-<b>Signature:</b>
-
-```typescript
-outputTableDisabled?: boolean;
-```
