--- javascript/react/docs/markdown/react.stepcomponentprops.outputtablelabel.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepComponentProps](./react.stepcomponentprops.md) &gt; [outputTableLabel](./react.stepcomponentprops.outputtablelabel.md)
-
-## StepComponentProps.outputTableLabel property
-
-The label to use for the output field
-
-<b>Signature:</b>
-
-```typescript
-outputTableLabel?: string;
-```
