--- javascript/react/docs/markdown/react.stepcomponentprops.step.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepComponentProps](./react.stepcomponentprops.md) &gt; [step](./react.stepcomponentprops.step.md)
-
-## StepComponentProps.step property
-
-The processing step
-
-<b>Signature:</b>
-
-```typescript
-step: Step;
-```
