--- javascript/react/docs/markdown/react.stepcomponentprops.workflow.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepComponentProps](./react.stepcomponentprops.md) &gt; [workflow](./react.stepcomponentprops.workflow.md)
-
-## StepComponentProps.workflow property
-
-The workflow instance
-
-<b>Signature:</b>
-
-```typescript
-workflow: Workflow;
-```
