+++ javascript/react/docs/markdown/react.stepform.md
@@ -1,13 +1,13 @@
 <!-- Do not edit this file. It is automatically generated by API Documenter. -->
 
-[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepComponent](./react.stepcomponent.md)
+[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepForm](./react.stepform.md)
 
-## StepComponent variable
+## StepForm variable
 
 Let's us render the Steps in a loop while memoing all the functions
 
 <b>Signature:</b>
 
 ```typescript
-StepComponent: React.FC<StepComponentProps>
+StepForm: React.FC<StepFormProps>
 ```
