+++ javascript/react/docs/markdown/react.stepformprops.hideinput.md
@@ -1,8 +1,8 @@
 <!-- Do not edit this file. It is automatically generated by API Documenter. -->
 
-[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepComponentProps](./react.stepcomponentprops.md) &gt; [hideInput](./react.stepcomponentprops.hideinput.md)
+[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepFormProps](./react.stepformprops.md) &gt; [hideInput](./react.stepformprops.hideinput.md)
 
-## StepComponentProps.hideInput property
+## StepFormProps.hideInput property
 
 Indicates that the input table should be hidden or else shown and editable by the user. It may be desirable to hide this if the modal is launched directly from a table, which would make display redundant.
