+++ javascript/react/docs/markdown/react.stepformprops.hideinputcolumn.md
@@ -1,8 +1,8 @@
 <!-- Do not edit this file. It is automatically generated by API Documenter. -->
 
-[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepComponentProps](./react.stepcomponentprops.md) &gt; [hideInputColumn](./react.stepcomponentprops.hideinputcolumn.md)
+[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepFormProps](./react.stepformprops.md) &gt; [hideInputColumn](./react.stepformprops.hideinputcolumn.md)
 
-## StepComponentProps.hideInputColumn property
+## StepFormProps.hideInputColumn property
 
 Indicates that the input column should be hidden or else shown and editable by the user. It may be desirable to hide this if the transform is expected to use the selected column when the modal is launched directly from a table.
