+++ javascript/react/docs/markdown/react.stepformprops.hideoutput.md
@@ -1,8 +1,8 @@
 <!-- Do not edit this file. It is automatically generated by API Documenter. -->
 
-[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepComponentProps](./react.stepcomponentprops.md) &gt; [hideOutput](./react.stepcomponentprops.hideoutput.md)
+[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepFormProps](./react.stepformprops.md) &gt; [hideOutput](./react.stepformprops.hideoutput.md)
 
-## StepComponentProps.hideOutput property
+## StepFormProps.hideOutput property
 
 Indicates that the output table should be hidden or else shown and editable by the user. It may be desirable to hide this if the transform is expected to do an inline replacement of the input table.
