+++ javascript/react/docs/markdown/react.stepformprops.index.md
@@ -1,0 +1,13 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepFormProps](./react.stepformprops.md) &gt; [index](./react.stepformprops.index.md)
+
+## StepFormProps.index property
+
+The step index
+
+<b>Signature:</b>
+
+```typescript
+index: number;
+```
