+++ javascript/react/docs/markdown/react.stepformprops.inputcolumnlabel.md
@@ -1,0 +1,13 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepFormProps](./react.stepformprops.md) &gt; [inputColumnLabel](./react.stepformprops.inputcolumnlabel.md)
+
+## StepFormProps.inputColumnLabel property
+
+Label for the input table dropdown
+
+<b>Signature:</b>
+
+```typescript
+inputColumnLabel?: string;
+```
