+++ javascript/react/docs/markdown/react.stepformprops.md
@@ -1,0 +1,32 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepFormProps](./react.stepformprops.md)
+
+## StepFormProps interface
+
+<b>Signature:</b>
+
+```typescript
+export interface StepFormProps 
+```
+
+## Properties
+
+|  Property | Modifiers | Type | Description |
+|  --- | --- | --- | --- |
+|  [hideInput?](./react.stepformprops.hideinput.md) |  | boolean | <i>(Optional)</i> Indicates that the input table should be hidden or else shown and editable by the user. It may be desirable to hide this if the modal is launched directly from a table, which would make display redundant. |
+|  [hideInputColumn?](./react.stepformprops.hideinputcolumn.md) |  | boolean | <i>(Optional)</i> Indicates that the input column should be hidden or else shown and editable by the user. It may be desirable to hide this if the transform is expected to use the selected column when the modal is launched directly from a table. |
+|  [hideOutput?](./react.stepformprops.hideoutput.md) |  | boolean | <i>(Optional)</i> Indicates that the output table should be hidden or else shown and editable by the user. It may be desirable to hide this if the transform is expected to do an inline replacement of the input table. |
+|  [index](./react.stepformprops.index.md) |  | number | The step index |
+|  [inputColumnLabel?](./react.stepformprops.inputcolumnlabel.md) |  | string | <i>(Optional)</i> Label for the input table dropdown |
+|  [inputTableLabel?](./react.stepformprops.inputtablelabel.md) |  | string | <i>(Optional)</i> Label for the input-table dropdown |
+|  [metadata?](./react.stepformprops.metadata.md) |  | TableMetadata | <i>(Optional)</i> TableMetadata |
+|  [onChange](./react.stepformprops.onchange.md) |  | (step: Step, index: number) =&gt; void | The step onchange handler |
+|  [onChangeOutput](./react.stepformprops.onchangeoutput.md) |  | (value: string \| undefined) =&gt; void | Event handler for when the output table name changes |
+|  [output?](./react.stepformprops.output.md) |  | string | <i>(Optional)</i> The output table name |
+|  [outputColumnLabel?](./react.stepformprops.outputcolumnlabel.md) |  | string | <i>(Optional)</i> The label to use for the output column field |
+|  [outputTableDisabled?](./react.stepformprops.outputtabledisabled.md) |  | boolean | <i>(Optional)</i> Whether the output table name is disabled |
+|  [outputTableLabel?](./react.stepformprops.outputtablelabel.md) |  | string | <i>(Optional)</i> The label to use for the output field |
+|  [step](./react.stepformprops.step.md) |  | Step | The processing step |
+|  [workflow](./react.stepformprops.workflow.md) |  | Workflow | The workflow instance |
+
