+++ javascript/react/docs/markdown/react.stepformprops.metadata.md
@@ -1,0 +1,13 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepFormProps](./react.stepformprops.md) &gt; [metadata](./react.stepformprops.metadata.md)
+
+## StepFormProps.metadata property
+
+TableMetadata
+
+<b>Signature:</b>
+
+```typescript
+metadata?: TableMetadata;
+```
