+++ javascript/react/docs/markdown/react.stepformprops.onchange.md
@@ -1,0 +1,13 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepFormProps](./react.stepformprops.md) &gt; [onChange](./react.stepformprops.onchange.md)
+
+## StepFormProps.onChange property
+
+The step onchange handler
+
+<b>Signature:</b>
+
+```typescript
+onChange: (step: Step, index: number) => void;
+```
