+++ javascript/react/docs/markdown/react.stepformprops.onchangeoutput.md
@@ -1,8 +1,8 @@
 <!-- Do not edit this file. It is automatically generated by API Documenter. -->
 
-[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepComponentProps](./react.stepcomponentprops.md) &gt; [onChangeOutput](./react.stepcomponentprops.onchangeoutput.md)
+[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepFormProps](./react.stepformprops.md) &gt; [onChangeOutput](./react.stepformprops.onchangeoutput.md)
 
-## StepComponentProps.onChangeOutput property
+## StepFormProps.onChangeOutput property
 
 Event handler for when the output table name changes
