+++ javascript/react/docs/markdown/react.stepformprops.output.md
@@ -1,0 +1,13 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepFormProps](./react.stepformprops.md) &gt; [output](./react.stepformprops.output.md)
+
+## StepFormProps.output property
+
+The output table name
+
+<b>Signature:</b>
+
+```typescript
+output?: string;
+```
