+++ javascript/react/docs/markdown/react.stepformprops.outputcolumnlabel.md
@@ -1,0 +1,13 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepFormProps](./react.stepformprops.md) &gt; [outputColumnLabel](./react.stepformprops.outputcolumnlabel.md)
+
+## StepFormProps.outputColumnLabel property
+
+The label to use for the output column field
+
+<b>Signature:</b>
+
+```typescript
+outputColumnLabel?: string;
+```
