+++ javascript/react/docs/markdown/react.stepformprops.outputtabledisabled.md
@@ -1,0 +1,13 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepFormProps](./react.stepformprops.md) &gt; [outputTableDisabled](./react.stepformprops.outputtabledisabled.md)
+
+## StepFormProps.outputTableDisabled property
+
+Whether the output table name is disabled
+
+<b>Signature:</b>
+
+```typescript
+outputTableDisabled?: boolean;
+```
