+++ javascript/react/docs/markdown/react.stepformprops.outputtablelabel.md
@@ -1,0 +1,13 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepFormProps](./react.stepformprops.md) &gt; [outputTableLabel](./react.stepformprops.outputtablelabel.md)
+
+## StepFormProps.outputTableLabel property
+
+The label to use for the output field
+
+<b>Signature:</b>
+
+```typescript
+outputTableLabel?: string;
+```
