+++ javascript/react/docs/markdown/react.stepformprops.step.md
@@ -1,0 +1,13 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepFormProps](./react.stepformprops.md) &gt; [step](./react.stepformprops.step.md)
+
+## StepFormProps.step property
+
+The processing step
+
+<b>Signature:</b>
+
+```typescript
+step: Step;
+```
