+++ javascript/react/docs/markdown/react.stepformprops.workflow.md
@@ -1,0 +1,13 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@datashaper/react](./react.md) &gt; [StepFormProps](./react.stepformprops.md) &gt; [workflow](./react.stepformprops.workflow.md)
+
+## StepFormProps.workflow property
+
+The workflow instance
+
+<b>Signature:</b>
+
+```typescript
+workflow: Workflow;
+```
