+++ javascript/react/src/components/StepForm/StepForm.types.ts
@@ -5,7 +5,7 @@
 import type { TableMetadata } from '@datashaper/tables'
 import type { Step, Workflow } from '@datashaper/workflow'
 
-export interface StepComponentProps {
+export interface StepFormProps {
 	/**
 	 * The workflow instance
 	 */
