+++ javascript/react/src/components/StepForm/index.ts
@@ -2,5 +2,5 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-export { StepComponent } from './StepComponent.js'
-export type { StepComponentProps } from './StepComponent.types.js'
+export { StepForm } from './StepForm.js'
+export type { StepFormProps } from './StepForm.types.js'
