+++ javascript/react/src/components/index.ts
@@ -6,8 +6,8 @@ export * from './ArqueroDetailsList/index.js'
 export * from './ArqueroTableHeader/index.js'
 export * from './Guidance/index.js'
 export * from './StepCard/index.js'
-export * from './StepComponent/index.js'
 export * from './StepDescription/index.js'
+export * from './StepForm/index.js'
 export * from './StepHistoryList/index.js'
 export * from './StepList/index.js'
 export * from './StepSelector/index.js'
